% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIXTCOMP_getParam.R
\name{getParam}
\alias{getParam}
\title{Get the estimated parameter}
\usage{
getParam(outMixtComp, var)
}
\arguments{
\item{outMixtComp}{object of class \emph{MixtCompLearn} or \emph{MixtComp} obtained using \link{mixtCompLearn} or \link{mixtCompPredict} functions.}

\item{var}{name of the variable to get parameter}
}
\value{
the parameter of the variable
}
\description{
Get the estimated parameter
}
\examples{
\donttest{
data(simData)
 
# define the algorithm's parameters
algo <- list(nbBurnInIter = 100,
             nbIter = 100,
             nbGibbsBurnInIter = 50,
             nbGibbsIter = 50,
             nInitPerClass = 10,
             nSemTry = 20,
             confidenceLevel = 0.95)

# run RMixtCompt for clustering
resLearn <- mixtCompLearn(simData$dataLearn$matrix, simData$model$unsupervised, algo, nClass = 2)

# get param
param <- getParam(resLearn, "Gaussian1")
}

}
\seealso{
\code{\link{plotDataBoxplot}}  \code{\link{plotDataCI}}

Other getter: \code{\link{getBIC}},
  \code{\link{getCompletedData}},
  \code{\link{getEmpiricTik}}, \code{\link{getPartition}},
  \code{\link{getType}}
}
\author{
Quentin Grimonprez
}
\concept{getter}
