% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JsonMixtComp.R
\name{JMixtCompLearn}
\alias{JMixtCompLearn}
\alias{JMixtCompPredict}
\title{Learn and predict using RJMixtComp}
\usage{
JMixtCompLearn(data, model, algo = createAlgo(), nClass, inputPath, outputFile)

JMixtCompPredict(
  data,
  model,
  algo = createAlgo(),
  nClass,
  inputPath,
  paramFile,
  outputFile
)
}
\arguments{
\item{data}{a data.frame/matrix/list containing the data}

\item{model}{a list containing the model for each variable. The names of \emph{model}'s element must be in \emph{data}}

\item{algo}{a list containing the parameters of the SEM-Gibbs algorithm (see \link{createAlgo}).}

\item{nClass}{the number of class of the mixture model.}

\item{inputPath}{path of the folder to save the input files.}

\item{outputFile}{path of the output json file to save.}

\item{paramFile}{(only for \emph{JMixtCompPredict}) path of the output json file of \emph{JMixtCompLearn} function.}
}
\value{
A json file and a R list containing 3 lists :
\describe{
 \item{algo}{a copy of \emph{algo} parameter.}
 \item{mixture}{information about the mixture (see \emph{Details}).}
 \item{variable}{information about the estimated parameters (see \emph{Details}).}
}
}
\description{
Estimate the parameter of a mixture model or predict the cluster of new samples.
}
\details{
See external documentation for details about the input/output object of \emph{JMixtCompLearn} and \emph{JMixtCompPredict} functions.
}
\examples{
\donttest{
library(jsonlite)

pathToData <- system.file("extdata", "data.json", package = "RJMixtComp")
pathToModel <- system.file("extdata", "desc.json", package = "RJMixtComp")

data <- as.data.frame(fromJSON(pathToData))
model <- fromJSON(pathToModel)

algo <- createAlgo()

resLearn <- JMixtCompLearn(data, model, algo = algo, nClass = 2,
                           inputPath = ".", outputFile = "reslearn.json")


file.remove("./algo.json", "./model.json", "./data.json")

resPredict <- JMixtCompPredict(data, model, algo = algo, nClass = 2,
                                  inputPath = ".", paramFile = "reslearn.json",
                                  outputFile = "respredict.json")


file.remove("./algo.json", "./model.json", "./data.json", "reslearn.json",
            "respredict.json")
}

}
\author{
Quentin Grimonprez
}
