% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIXTCOMP_getter.R
\name{getCompletedData}
\alias{getCompletedData}
\title{Get the completed data from MixtComp object}
\usage{
getCompletedData(outMixtComp, with.z_class = FALSE)
}
\arguments{
\item{outMixtComp}{object of class \emph{MixtCompLearn} or \emph{MixtComp} obtained using \link{mixtCompLearn} or \link{mixtCompPredict} functions.}

\item{with.z_class}{if TRUE, z_class is returned with the data.}
}
\value{
a matrix with the data completed by MixtComp (z_class is in the first column and then variables are sorted in alphabetic order, it may differ from the original order of the data).
}
\description{
Get the completed data from MixtComp object
}
\examples{
\donttest{
data(simData)
 
# define the algorithm's parameters
algo <- list(nbBurnInIter = 100,
             nbIter = 100,
             nbGibbsBurnInIter = 50,
             nbGibbsIter = 50,
             nInitPerClass = 10,
             nSemTry = 20,
             confidenceLevel = 0.95)

#' # keep only 3 variables
model <- simData$model$unsupervised[c("Gaussian1", "Poisson1", "Categorical1")]

# run RMixtCompt for clustering
resLearn <- mixtCompLearn(simData$dataLearn$matrix, model, algo, nClass = 2)

# get completedData
getCompletedData <- completed(resLearn)
}

}
\seealso{
Other getter: \code{\link{getBIC}},
  \code{\link{getEmpiricTik}}, \code{\link{getParam}},
  \code{\link{getPartition}}, \code{\link{getType}}
}
\author{
Quentin Grimonprez
}
\concept{getter}
