% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageDescription.R
\docType{package}
\name{RJMixtComp-package}
\alias{RJMixtComp-package}
\title{RJMixtComp}
\description{
Mixture Composer is a project to build mixture models with heterogeneous data sets. The conditional independence property enable composing all the data sets in a single mixture model.
}
\details{
Use JMixtComp library.
}
\examples{
\donttest{
pathToData <- system.file("extdata", "data.json", package = "RJMixtComp")
pathToModel <- system.file("extdata", "desc.json", package = "RJMixtComp")

data <- as.data.frame(fromJSON(pathToData))
model <- fromJSON(pathToModel)
algo <- createAlgo()

resLearn <- JMixtCompLearn(data, model, algo = algo, nClass = 2,
                           inputPath = ".", outputFile = "reslearn.json")


file.remove("./algo.json", "./model.json", "./data.json")

resPredict <- JMixtCompPredict(data, model, algo = algo, nClass = 2, inputPath = ".",
                               paramFile = "reslearn.json", outputFile = "respredict.json")


file.remove("./algo.json", "./model.json", "./data.json", "reslearn.json",
            "respredict.json")
}


}
\author{
For RJMixtComp: Quentin Grimonprez

Maintainer: Quentin Grimonprez <quentin.grimonprez@inria.fr>

For JMixtComp:
Vincent Kubicki <vincent.kubicki@inria.fr">
Christophe Biernacki
Quentin Grimonprez
Serge Iovleff
}
\keyword{package}
